backend/Mongodb/requirements.txt
fastapi
pymongo
uvicorn
requests
grpcio-tools
grpcio
